Regional reports added

+++ scripts/main.py
@@ -15,87 +15,100 @@ from report import Report
 import shutil
 import charts
 from queries_general import run_general_json_generation
-
-
+from general_report import run_general_report
 
 
 start_time = time.time()
+
+multiprocessing.set_start_method("spawn", force=True)
+
 dotenv_path = Path(__file__).parent / ".env"
 load_dotenv(dotenv_path=dotenv_path)
 
-# Definir credenciales de conexión manualmente
 connect_args = {
-    "url": os.getenv("AXONIUS_URL"),  # IP de la instancia de Axonius
-    "key": os.getenv("AXONIUS_KEY"),  # API Key proporcionada
-    "secret": os.getenv("AXONIUS_SECRET"),  # API Secret proporcionado
-    "verify": False  # Desactivar la verificación SSL (No recomendado en producción)
+    "url": os.getenv("AXONIUS_URL"),
+    "key": os.getenv("AXONIUS_KEY"),
+    "secret": os.getenv("AXONIUS_SECRET"),
+    "verify": False
 }
 
 current_date_and_time = str(dt.date.today())
 
 
 
-
 def run_axonius(central):
-    #Ejecuta todas las consultas de Axonius en paralelo
     processes = []
-    for x in range(len(central.queries)):
-        p = multiprocessing.Process(target=axonius_retreive_data, kwargs={
-            "connect_args": connect_args,
-            "saved_query_name": central.queries[x],
-            "saved_query_name_clean": central.file_name[x],
-            "central": central.nombre,
-            "current_date_and_time": current_date_and_time,
-        })
+
+    for query, filename in zip(central.queries, central.file_name):
+
+        while len(processes) >= 4:
+            for p in processes:
+                if not p.is_alive():
+                    processes.remove(p)
+            time.sleep(0.5)
+
+        p = multiprocessing.Process(
+            target=axonius_retreive_data,
+            kwargs={
+                "connect_args": connect_args,
+                "saved_query_name": query,
+                "saved_query_name_clean": filename,
+                "central": central.nombre,
+                "current_date_and_time": current_date_and_time,
+            }
+        )
+
         processes.append(p)
         p.start()
 
     for p in processes:
         p.join()
+
 
 def run_critical(central):
-    #Ejecuta la función critical en paralelo para las severidades CRITICAL y HIGH
     processes = []
+
     for severity in ["CRITICAL", "HIGH"]:
-        p = multiprocessing.Process(target=get_severities, kwargs={
-            "central": central.nombre,
-            "f_central": central.fullName,
-            "severidad": severity,
-        })
+        p = multiprocessing.Process(
+            target=get_severities,
+            kwargs={
+                "central": central.nombre,
+                "f_central": central.fullName,
+                "severidad": severity,
+            }
+        )
+
         processes.append(p)
         p.start()
 
     for p in processes:
         p.join()
 
-def run_eol(central):
-    #Ejecuta la función critical en paralelo para las severidades CRITICAL y HIGH
-    processes = []
-    p = multiprocessing.Process(target=export_eol, kwargs={
-        "central": central.nombre,
-        "f_central": central.fullName,
-    })
-    processes.append(p)
-    p.start()
 
-    for p in processes:
-        p.join()
+def run_eol(central):
+    p = multiprocessing.Process(
+        target=export_eol,
+        kwargs={
+            "central": central.nombre,
+            "f_central": central.fullName,
+        }
+    )
+
+    p.start()
+    p.join()
 
 
 def run_reporte(central):
-    processes = []
     p = multiprocessing.Process(target=Report, args=(central,))
-    processes.append(p)
     p.start()
+    p.join()
 
-    for p in processes:
-        p.join()
+
 
 def only_ixtla_carso_or_both(centrales):
     nombres = {c.nombre for c in centrales}
-    return (
-        nombres == {"IXTLA"}
-    )
+    return nombres == {"IXTLA"}
+
 
 def general_json_done_exists():
     today = str(dt.date.today())
@@ -103,53 +116,53 @@ def general_json_done_exists():
     done_file = base_dir / f"general_json_{today}.done"
     return done_file.exists()
 
+
+
 if __name__ == '__main__':
 
     if not only_ixtla_carso_or_both(centrales):
         if not general_json_done_exists():
-           run_general_json_generation(max_workers=5, delete_previous=True)
-    processes = []
+            run_general_json_generation(max_workers=5, delete_previous=True)
+
+    GENERAL_REPORT_CENTRALS = {f"R{i}" for i in range(1, 10)} | {"GENERAL"}
+
     for central in centrales:
-        path = r'./ARCHIVOS_REPORTES/'+central.nombre
-        if not os.path.exists(path):
-            os.mkdir(path)
-        path = path + r'/'+current_date_and_time
-        if not os.path.exists(path):
-            os.mkdir(path)
-        path_done = path + r'/done'
-        if not os.path.exists(path_done):
-            os.mkdir(path_done)
-
-        # Crear procesos para ejecutar Axonius y Critical en paralelo
-        p1 = multiprocessing.Process(target=run_axonius, args=(central,))
-        p2 = multiprocessing.Process(target=run_critical, args=(central,))
-        p3 = multiprocessing.Process(target=run_eol,args=(central,))
-        p4 = multiprocessing.Process(target=run_reporte,args=(central,))
-        processes.extend([p1, p2, p3, p4])
+        path = f'./ARCHIVOS_REPORTES/{central.nombre}/{current_date_and_time}/done'
+        os.makedirs(path, exist_ok=True)
+        if central.nombre in GENERAL_REPORT_CENTRALS:
+            run_general_report(central.nombre)
+        
+        else:
+            run_axonius(central)
+            run_critical(central)
+            run_eol(central)
+            run_reporte(central)
 
 
-        # Iniciar ambos procesos
-        p1.start()
-        p2.start()
-        p3.start()
-        p4.start()
-
-    # Esperar a que todos los procesos terminen
-    for p in processes:
-        p.join()
     time.sleep(5)
-    for central in centrales:
-        try:
-            shutil.rmtree(f'./ARCHIVOS_REPORTES/{central.nombre}/{current_date_and_time}/done')
-        except Exception as e:
-            print(f"⚠️ Error al eliminar {path}: {e}")
-    for central in centrales:
-        try:
-            charts.data_central(central.nombre)
-            charts.agregar_hojas_graficas(central.nombre)
 
-        except Exception as e:
-            print(f"Error : {e}")
-    end_time = time.time()  # Captura el tiempo de finalización
-    elapsed_time = (end_time - start_time) / 60  # Calcula el tiempo transcurrido
-    print(f"⏳ Tiempo de ejecución: {elapsed_time:.2f} minutos")
+    for central in centrales:
+        if central.nombre in GENERAL_REPORT_CENTRALS:
+            run_general_report(central.nombre)
+        
+        else:
+            try:
+                shutil.rmtree(f'./ARCHIVOS_REPORTES/{central.nombre}/{current_date_and_time}/done')
+            except Exception as e:
+                print(f"⚠️ Error al eliminar carpeta done: {e}")
+
+    for central in centrales:
+        if central.nombre in GENERAL_REPORT_CENTRALS:
+            run_general_report(central.nombre)
+        
+        else:
+            try:
+                charts.data_central(central.nombre)
+                charts.agregar_hojas_graficas(central.nombre)
+            except Exception as e:
+                print(f"Error generando gráficas: {e}")
+
+    end_time = time.time()
+    elapsed_time = (end_time - start_time) / 60
+
+    print(f"\nTiempo de ejecución: {elapsed_time:.2f} minutos")
